Responsibility refactor and test declaration

+++ src/main/java/com/bookstore/persistence/controller/BookController.java
@@ -2,6 +2,7 @@ package com.bookstore.persistence.controller;
 
 import java.util.List;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -14,62 +15,61 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.bookstore.persistence.dto.BookDTO;
-import com.bookstore.persistence.model.Book;
 import com.bookstore.persistence.service.BookService;
 
 @RestController
 @RequestMapping("/bookstore/api/dao/books")
 public class BookController {
 
+	@Autowired
 	private final BookService bookService;
 
 	public BookController(BookService bookService) {
 		this.bookService = bookService;
 	}
 
-	@PostMapping
-	public ResponseEntity<BookDTO> createBook(@RequestBody Book book) {
-		BookDTO createdBook = bookService.createBook(book);
-		return ResponseEntity.status(HttpStatus.CREATED).body(createdBook);
-	}
-
 	@GetMapping
-	public ResponseEntity<Object> getAllBooks() {
+	public ResponseEntity<List<BookDTO>> getAllBooks() {
 		List<BookDTO> books = bookService.getAllBooks();
-		if (books != null) {
-			return ResponseEntity.ok(books);
-		} else {
-			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("No books found");
-		}
+		return new ResponseEntity<>(books, HttpStatus.OK);
 	}
 
 	@GetMapping("/{id}")
-	public ResponseEntity<Object> getBookById(@PathVariable Long id) {
+	public ResponseEntity<BookDTO> getBookById(@PathVariable Long id) {
 		BookDTO book = bookService.getBookById(id);
-		if (book != null) {
-			return ResponseEntity.ok(book);
-		} else {
-			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("No books found with id: "+id);
+		if (book == null) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
+		return new ResponseEntity<>(book, HttpStatus.OK);
+	}
+
+	@PostMapping
+	public ResponseEntity<BookDTO> addBook(@RequestBody BookDTO bookDTO) {
+		BookDTO createdBook = bookService.addBook(bookDTO);
+		return new ResponseEntity<>(createdBook, HttpStatus.CREATED);
 	}
 
 	@PutMapping("/{id}")
-	public ResponseEntity<BookDTO> updateBook(@PathVariable Long id, @RequestBody Book updatedBook) {
-		BookDTO book = bookService.updateBook(id, updatedBook);
-		if (book != null) {
-			return ResponseEntity.ok(book);
-		} else {
-			return ResponseEntity.notFound().build();
+	public ResponseEntity<BookDTO> updateBook(@PathVariable Long id, @RequestBody BookDTO bookDTO) {
+		if (!bookService.bookExists(id)) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
+
+		if (!bookDTO.getId().equals(id)) {
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		}
+
+		BookDTO updatedBook = bookService.updateBook(bookDTO);
+		return new ResponseEntity<>(updatedBook, HttpStatus.OK);
 	}
 
 	@DeleteMapping("/{id}")
 	public ResponseEntity<Void> deleteBook(@PathVariable Long id) {
-		boolean deleted = bookService.deleteBook(id);
-		if (deleted) {
-			return ResponseEntity.noContent().build();
-		} else {
-			return ResponseEntity.notFound().build();
+		if (!bookService.bookExists(id)) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
+
+		bookService.deleteBook(id);
+		return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 	}
 }
